// types/tutor.ts
export type Tutor = {
    _id: string;
    name: string;
    subjects: string[];
    experience: string;
    qualification: string;
    timing: string;
    areas: string[];
    rating: number;
    classes: string[];
    boards: string[];
    availability: string;
    photo: string;
    bio: string;
  };